pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "informatica-to-dbt"
version = "1.0.0"
description = "Production-grade Informatica PowerCenter XML to dbt project converter for Snowflake"
requires-python = ">=3.9"
dependencies = [
    "lxml>=4.9",
    "networkx>=3.0",
    "sqlparse>=0.4",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.scripts]
infa2dbt = "informatica_to_dbt.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
